Tag In Space
AI Scripts for Space Tag where one ship hunts while the rest try to run.
The player is included.
When the AI run, they don't run directly away because that would be
boring.  Instead they randomly decide to follow one of a few directions.

Unfortunately, the only way to tell what's going on is to turn on the
in-game console, which cut my FPS to a third of normal.

git-svn-id: http://epiar.net/repo/epiar@76 40d08037-4d43-4175-bee4-ccfa03fdb620