Cleaned up the TAG game
Created two global tables: swarm and it to make things more packaged.
Pulled the decision making into a set of stand alone function 'plans'.
Each ship has their own plan, and since the plans are partially state
based, each ship moves unpredictably.  This plan model could be adapted
to work for much more complex

Added a Window that gives live feedback on which ship is IT.
This window reflects the rotation of the IT ship.

Unlike the earlier version, users should be able to figure out what's
going on now.  Like the earlier version, the AI do not do the smart
thing and just flee, since that would be boring.