Fixed the 'Store'.
The Store buttons that were not on the top row were being clicked on
because they were being overlapped by the ship Pictures.

This cleans up the Store layout.
Adds a commented out debug tool to see the 'real' and 'display' areas of
Pictures and the Images they hold.
Adds some debug to various Widgets.