Use the QuadTree GetSpritesNear in the HUD Radar.
This shrinks the Radar visibility in order to make the effect more
pronounced.  Eventually, I think that Radar visibility should be set in
option.xml or modified by different Player Equiptment.  The smaller
Radar makes it more obvious that all blips are circles.

Adds the Player to the Radar with a yellow color just fun.