Creation of Animation Tool
Because of the custom format, the .ani files were difficult to view or
edit.  This python tool provides a way to pack and unpack Animations.

The tool converts .ani files into folders and vise-versa.
To unpack all Epiar animations, run:
	Source/tools/ani.py Resources/Animations/*
To pack these files back up after editing:
	Source/tools/ani.py *_unpacked
	mv -f *.ani Resources/Animations

Creating new .ani files requires putting png files and running ani.py on
it.  Files are ordered alphabetically unless an order.txt file is
included.  The frame length can be defined in delay.txt or with the
--delay option.  If no delay is provided, the user is prompted.