Moves all Player actions to Registered Key Mappings
The keyDown list was badly named (easily confused with KEYDOWN) so it was renamed heldKeys.

The heldKeys list is pushed to the end of the events list as KEYPRESSED events.  KEYPRESSED events will trigger for as long as that key remains pressed down.

The Lua Key mappings are accpeting numbers now rather than characters.  I don't know how to accept either type.