Adds Optional focusing on Non-Player Sprites.
Everything should be drawn around the Sprite that has Camera Focus.
SpriteManager and Hud now reference the Camera Focus rather than the Player.

The keys 'q' and 'w' change focus between the player and the player's target.

This feature should be turned off in the normal game and only used during debug.