Fixes issues with Keyboard Input
Since the SDLKey is an enum rather than a char, truncating the SDLKey means that some characters (notably all the meta-characters like shift and alt) were being silently converted to normal characters.

This solves a problem with meta characters like shift not generating a KEYTYPED event.  Now Meta-characters are typed but the Textbox and Console just ignore them.

This adds character support for more keys.

This still makes the assumption that users have an American Keyboard.  This will need to be fixed.