Splits Player Mode and Debug Mode
In debug mode, the user does not pilot a ship.  Instead the player pans the camera. (The speed is exponential and tied to the radar visibility.)

This creates a file commands.lua for managing Key Commands. This includes customizing Keys at runtime.  This allows us to have some similar commands and some commands in different modes.

debugMode is a flag to the Simulation::Run.  It is an xml option now, but should become a command line option later.