Adds scripts to create a Mac OS X Application Bundle
This adds special checking to main to detect if the executable is being run from inside a MacOS Package.  This may not be the best way of doing this, but it works.

The mkbundle.sh uses the autotools build to create a working Epiar.app.

The cmake "make package" will create a .dmg with a working Epiar.app inside. (Maoserr wrote most of the CMake script and I debugged it)

Neither method properly handles any of the Libraries so this will likely only work on my exact computer.

The Info.plist is very incomplete, but it will run Epiar and display an icns.