Changes AI decision making to a simple State Machine
AI has a StateMachine attribute with the Lua function name.
AI has a state attribute which is an argument to the StateMachine.

This creates two StateMachines Trader and Hunter.
Trader picks a random planet, travels towards it, stops, then finds a new planet.
Hunter picks a nearby ship and chases it until it dies, then picks a new target.