Fixes Scrollbars for Tabs
The main problem was that the TAB doesn't have dimensions (height and width) when it is first created, and may not have dimensions when Widgets are added to it.  This means that the TAB has to Reset the scrollbars whenever a Child is added or it is attached to a Tab Container.

This also changes the Shipyard and Armory to vertically align their goods to show off the Scrollbar in Tabs.