Moves Engines from the Model to the Ship
This allows Ships with the same Model to have different Engines.  For example, now the player can buy new Engines.

Unfortunately, just like weapons, this means that sometimes the Engines don't seem to fit with the Ship they're attached to (Escape Pods with Ion Engines for example).  We can fix this later.

This also adds some old Epiar 0.5.0 art for use as the Engine Store Pictures.