Adds Gate exit points
Ships can enter the Gate from any angle, but leave from the bottom.
Since there is no cooldown time, the exiting ships must be offset outside of the ring.
There is currently no restriction on what kind of Sprite can be used as an exit, but making it another Gate makes the most sense.
Currently, all Gates are pairs, but that's not neccessary in the long run.

This also changes the minimum Sprite ID to 1 and reserves Sprite ID 0 to nonexistent Sprites.