Adds a Outfit summary to each ship called shipStats
When acting, each ship uses it's own shipStats rather than the stats from the Engine, Model, or Weapon.

The shipStats are recalculated whenever the Engine or Model changes or a new Weapon is added.

This is done so that we can have stat bonuses without needing to recalculate the sum on every action.