Groups the random generation of Planets, Stations and Gates into Systems
This replaces the completely random scattering of Planets, Stations, and Gates.
This makes the gate more grid-like and helps to ensure that it's possible to get to any planet via a nearby gate.

The system layout is random, so loading a saved game is likely to drop the player in the middle of nowhere.

The planets and systems are named after their system ("Mosalia 3" or "Chenal Outpost 1").  The system names are hardcoded and there are only as many systems as there are names.

All the system names were created using a random name generator on the web, feel free to add or remove them.