Small UI fixes
Center text in the middle of the Textbox.
Compensate for the text when calculating Checkbox height and width.
Check for Focus in UIContainers by walking backwards through the children list; the last Widget drawn is the first Widget checked.
Add small pad to scrolling Tabs to prevent Widgets from being cut.