Change the way that XML is parsed into Components
This removes the parserCB and adds a FromXMLNode to each Component.

By giving the Component direct access to the Node, it can have more interesting structures than just key-value pairs.

For example, Cargo has both type and amount.  Previously, the type was being used as the node name, but spaces in the Cargo-Node names was invalid XML causing the Player saved game file to be corrupted.

This also gives the Component the ability to raise errors when an attribute is missing (eg, no X-Y position).

Unfortunately, the one-to-many relationships (eg, one technology has many models) are a bit more cumbersome since it now lacks the implicite looping and must be done explicitely.

There are two potential ideas that have not been explored:
 1) Multiple component types within a single xml file (eg, 'Planets & Gates' or 'Weapons & Ammo').
 2) Node overloading. Models, Engines, Outfits, and Weapons could all use the Outfit FromXMLNode for parsing overlapping attributes.