Fix bug when saving in the key settings
If the player opens up the option window, then load their Player, they cannot save or close the options window.  This changes the way that keys are processed: only the keys that were present when the command window was created are saved.

This also:
Increases the width of the Key Commands.
Turns off debug print statements.