Change the Camera focusSprite to a spriteID.
The Camera now needs to be passed the SpriteManager during update.
The SpriteManager no longer has to compare on every sprite deletetion.

This makes the Camera follow the general "No Storage of Sprite Pointers".

The camera now doesn't have any Global dependencies, but since it is still being used as a Singleton, we can't make it a non-singleton yet.