Use the Planet's current location when loading the player.
This allows the planet's location to change while still allowing the player to load to a valid location.  This could happen if the user is edits the universe since the last save, or if the player is using the random-universe option.

This also saves the player when they land and when they leave a planet so that the purchases made on that planet are recorded.