Fix projectile momentum bug.
The lastUpdateFrame for new sprites was being initialized to zero, which
meant that on the first frame that the sprites updated, it will move by
momentum*( the total number of logical frames so far).  For most
Sprites, this isn't a problem since their starting momentum is zero.
But for the Projectile, they start with the momentum of the ship that
fired them.  This nonzero momentum was multiplied thousands of times and
moved the projectile so far past the target as to completely miss them.

This reveals another problem with Projectiles: they only hit things that
are near them when they update.  If the projectile is moving quickly, it
can pass through ships without colliding.