Change the way that StatusBars change values
Before, there was a function "UpdateHUD" that needed to be called every tick and would update each StatusBar.  Now each StatusBar has a "lua_updater" string that it uses to fetch it's new value on each update tick.

The lua_updater string is evaluated and the return value is used as the new name or ratio value.  The evaluation is implicitly returned, so the Lua script can use "7" or "foo()"rather than "return 7" or "return foo()".

This also moves the starfield, camera, and Hud Update calls outside of the pause check.  (This way the 'credits' StatusBar can update while the User is buying things.)