Move the Power Distribution keymap and window
The keymap is now:
Increase: (A)ccelerate, (S)hields, (D)amage.
Decrease: (Z),          (X),       (C)

(Unfortunately, (Z) is a bad idea for German keyboards.)

This is because (m) was the map and pretty vital at this point.

Moves the Power Management Window to the bottom center of the screen.
This is because it was overlapping weapon StatusBars.