Fix ugly StatusBar shield display
We count the amount of damage that the hull and shields have taken rather than how much damage they can still absorb.  This means that the damage values can be greater than the Max value.  For Hull this simply means that the ship is destroyed, but for Shields this was leading to a negative number.

The Hud StatusBars only use the ratio-bar display when a value is between 0.0 and 1.0.  When a StatusBar got a negative Shield or Hull Percentage, it would interpret that as a string rather than a ratio.  Normal Shield values weren't clearing the text field.  Since the StatusBars can theoretically have both string values and ratio values, the StatusBars were drawing the string and drawing the ratio.  This meant that even refocusing couldn't clear the StatusBar values.

Now:
Hull and Shield percentages will never return negative numbers.
Hud StatusBars set both values on every update; one is a real value while the other is blank.