Fix Window Constructor and Destructor errors.
In GCC, default arguments must be given in the order that they are listed in the class description or a warning is raised.

Window Scrollbars are either NULL or are added to the normal UIContainer children list.  Therefore, they cannot be deleted in the ~Window or they would be deleted twice.