Show Introduction Window every time a new player is created.
Don't show the Intro Window every time that the game is launched (unless the option is turned off) because this is really annoying.  Instead, show the Window whenever a new player is created.

This will make the game easier to demo for friends since it will explain the keyboard commands when they create new players, even if the more experience players already know the commands.

The OPTION was also a hassle since it would always be toggled during normal testing.  Causing it to be misrecognized as a change that should be pushed.