Turn clipping back on for UI::Button.
This clips text, an undesirable effect in player releases. But we
should not, ultimately, be drawing text outside the box. If text
doesn't fit, whoever is coding that dialog is responsible for its
bad design and must change it. The UI should not accept the mistakes
of other parts of the code.