Fixed (again) the landing behavior. Once to select, a second to land.
The basic problem here is if I have no planet selected and am stopped right
atop a planet, I should still have to press 'L' twice - once to select, a
second time to land. This way, the behavior is always consistent, regardless
of where you are. Inconsistent game behavior should be avoided as a rule.