Added verbose error logging to weapon XML loading & fix no sound issue.
Summary above is accurate but expanding on the no sound issue: weapon
XML loading would fail if the audio device hadn't initialized because
the sound files wouldn't load, which the system took to mean the XML
file didn't parse properly. We want Epiar to work if the audio device
is unavailable or disabled, so this removes that issue.