Fix build issues on MinGW and use pkg-config to search for libraries.
Note: You need to run ./autogen.sh and ./configure again after
applying this patch. If it fails to detect a library, please tell
Chris (topher) and feel free to use the old configure.in until
the issues are hammered out. Assumes Unix-esque install of some
libraries on OS X (Darwin ports or Macports), except for SDL,
SDL_image, and SDL_mixer which should be OS X frameworks as the
vendor offers them that way.