Use UI.search to check if a window exists rather than checking a global value.
The old method for preventing multiple copies of the same window from being opened was to check if the global lua instance that was used when creating the Window was still set.

if windowVar ~= nil then return end
windowVar = UI.newWindow( ... )

Now, the window is searched by name:

if UI.search("/'windowName'/") ~= nil then return end
UI.newWindow(... 'windowName')

While this is more verbose and computationally expensive, it allows the window to be manipulated by either Lua or C++.  I plan on removing as many global values as possible.

For example, allowing C++ to close the options window meant that the Lua variable optionWindow would not be invalidated and therefore Lua would not know that the options window had been closed.  If the user opened options with '?', hit ESC, then attempted to reopen the options window with '?', Lua would abort and the window would be forever closed.