Encapsulate the Store items in Frames
The "Items for Sale" list and the "Item on Display" are both in pretty frames.

Putting the list in a Frame prevents the height of the list from causing the entire tab to scroll.
Putting the item in a Frame provides a clean way of cleaning up Widgets when switching to a different Item.

Each time an item is selected, the old viewer is closed and a new viewer is created.  Since each Viewer is made specifically for the item it is showing, there are no longer any 'blank' attributes.

The entire store uses UI.search calls rather than global variables.