Lua Fleet class for abstracted control of grouped escorts
What it does is abstract the control of AIs into an easy-to-use
Fleet object. These objects are themeselves managed by Fleets,
which provides a wrapper to determine which fleet, if any, a
ship belongs to, or join one.

Current behavior includes:
  - Escorts are in "Accompanying" by default and can be told to return to that state
  - Leader can transmit a gate route to escorts and tell them to follow it
  - Leader can order escorts to attack a target

Freighter mission has been updated to demonstrate these features