Ticket #124 - Parse Command Line arguments before Initializing Singletons (like Video)
Init_Singletons has been split into two functions; options.xml and skin.xml files are loaded, then the arguments are parsed, then the rest of the Singletons are initialized.

Initializing the Video too early is a bad idea when the user supplies command line arguments that will stop the game from starting normally.  When the video initializes it creates a GUI window.