Two linewrap() improvements: hard wrap and justify
Justified text only works properly for fixed width fonts, but the
improvement is enough even for variable width fonts to make it
worthwhile. Better would be to have a C++ justify function which
is aware of individual glyph widths and can calculate the line
widths accordingly. FTGL may already support this.