Allow captured ships to be used as escorts
Matt suggested this feature. In order to make these escorts persist
across saves, there is a new player XML property hiredEscort, each
with two children, type and pay. Captured ships are "hired" for
0 credits. Actual paid escorts have not been implemented yet.

Note: Escorts created for missions are not considered "hired" and
so should never appear in this list.

To-do: It would probably make sense to place some limit on the number
of escorts a player can have (maybe 6). Also, there should be a way
to fully repair a captured escort other than saving and re-loading the
player.