Fix ConvertOldVersion to copy more than one Player
When copying the player's information to a new xml file, the original xmlnode would be modified so that the ->next attribute would be set to NULL.  This would mean that that Player would be loaded correctly, but all later players would be skipped.

This makes a copy of the player's node before creating the new xml file.