Give each Widget an array of Actions indexed by action_type
The Widget can register any Action to any of the action types, and Activate those functions whenever needed.  The Actions will Activate by default when that action type occurs (and isn't blocked by a subclass).

This allows the functionality of any Widget to be expanded easily for specific reactions.

An example usage:
  widget->RegisterAction( Action_MouseWDown, new LuaAction("print('Wheel Click!')") );
  widget->Activate(Action_MouseWDown);