Fix Bug #137 - Crash when Screens Swap when a Dropdown is open
The Screen Swap was not flushing the deferred list, so floating widgets were neither drawn nor forgotten.  Each time that one of the screens with an open dropdown, the drawwing of the Dropdown would be deffered, but not flushed.  This meant that there were several stale Dropdown pointers in the deferred list caused by Swapping screens.  If the UI is Closed before it is drawn, then those stale pointers will be accessed incorrectly.