Build system changes to allow 'make install' to work.
A later patch will add $(datadir)/epiar/ to the PhysicsFS search path.
Only then will Epiar actually work from a /usr/local/bin/epiar on a
random system. This is part of the work done to 1) get Epiar to work
easily and as expected for Linux distro packagers and 2) 'complete' our
autotools support by not ignoring that important last step.