Added .desktop file for Unix installations.
This file is in the root as it does not properly belong as part of
Resources/ - it isn't a file used by Epiar, but rather by the system
Epiar is installed on. README and NEWS are in much the same category,
and so I've placed them here.

Other projects (Freeciv, for instance), seems to place README and NEWS
in a doc/ and things like epiar.desktop in a bootstrap/. We could always
come up with our own conventions on this in the future.