Experimental Large Radar / Small Map
Click on the Radar to switch modes.  In the 'Large Radar Mode', a 300x300 Map Widget replaces the normal Map.  When the Radar visibility is below level 3 (4096 pixels radar radius), the Map will track the player exactly like the Radar.  At large visibilities, the Map can be panned and zoomed.

This is an experimental feature request and may be removed.