Change the way that video defaults are handled:
If full screen, overwrite defaults.
If zeros are being used for a width or height value, use height and width of the screen.

Width and Height should always be set by the default Options.

Before, the full screen was only being set if both width and height were zero, but the screen would still be stretched to the full screen.  This reduced the resolution without any benefit.