Changes to Universe flavor text
Using the current editor, I added a description to some of the game Components.

This creates a more planets and gates.  Some ship prices increased, but probably not by enough.

This removes all "Lorem Ipsum" fillers from the Planet descriptions.  The Planets need real summaries and this filler text takes too long to delete in the editor.