Add Jumping
This teleports a ship from one place to another.

The current jump animation is pretty bad and just moves the ship to the edge of the screen.

This is not set up correctly at this point: rather than making this a normal command hitting 'j' will jump to a random planet's sphere of influence.